Put command keyword to Util class, so we can refactor it easily.
