+++ RanBot/src/ranbot/command/CommandHandler.java
@@ -5,5 +5,5 @@ import sx.blah.discord.util.MissingPermissionsException;
 import sx.blah.discord.util.RateLimitException;
 
 public interface CommandHandler {
-	void processCommand(MessageCommand messageCommand) throws DiscordException, MissingPermissionsException, RateLimitException ;
+	void processCommand(MessageCommand messageCommand) throws DiscordException, MissingPermissionsException, RateLimitException;
 }
